drop down block

+++ src/main/java/com/waa/dragons/mediationattendance/domain/Student.java
@@ -18,11 +18,9 @@ import java.util.List;
 @Setter
 @NoArgsConstructor
 public class Student {
-
-
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
 
     private String firstName;
     private String lastName;
